#190 make text in steps bigger if enhanced accessibility enabled
