+++ app/src/main/java/com/github/braillesystems/learnbraille/ui/screens/theory/steps/AbstractStepFragment.kt
@@ -1,9 +1,13 @@
 package com.github.braillesystems.learnbraille.ui.screens.theory.steps
 
+import android.text.method.ScrollingMovementMethod
+import android.util.TypedValue
 import android.view.Menu
 import android.view.MenuInflater
 import android.view.MenuItem
 import android.widget.Button
+import android.widget.TextView
+import androidx.core.text.parseAsHtml
 import com.github.braillesystems.learnbraille.COURSE_ID
 import com.github.braillesystems.learnbraille.R
 import com.github.braillesystems.learnbraille.data.entities.Step
@@ -13,6 +17,7 @@ import com.github.braillesystems.learnbraille.ui.screens.HelpMsgId
 import com.github.braillesystems.learnbraille.ui.screens.theory.toCurrentStep
 import com.github.braillesystems.learnbraille.ui.screens.theory.toNextStep
 import com.github.braillesystems.learnbraille.ui.screens.theory.toPrevStep
+import com.github.braillesystems.learnbraille.utils.checkedAnnounce
 import com.github.braillesystems.learnbraille.utils.navigate
 import com.github.braillesystems.learnbraille.utils.setSize
 import org.koin.android.ext.android.inject
@@ -41,6 +46,19 @@ abstract class AbstractStepFragment(helpMsgId: HelpMsgId) : AbstractFragmentWith
             )
             nextButton?.setSize(
                 width = resources.getDimension(R.dimen.side_buttons_extended_width).toInt()
+            )
+        }
+    }
+
+    protected fun setText(text: String, infoTextView: TextView) {
+        infoTextView.text = text.parseAsHtml()
+        infoTextView.movementMethod = ScrollingMovementMethod()
+        checkedAnnounce(text)
+        if (preferenceRepository.extendedAccessibilityEnabled) {
+            // Size applied in runtime is different
+            infoTextView.setTextSize(
+                TypedValue.COMPLEX_UNIT_SP,
+                resources.getDimension(R.dimen.lessons_info_text_size) / 4 * 3
             )
         }
     }
