+++ app/src/main/res/values/dimens.xml
@@ -30,7 +30,6 @@
     <dimen name="lessons_info_width">270dp</dimen>
     <dimen name="lessons_info_height">0dp</dimen>
     <dimen name="lessons_info_text_size">18sp</dimen>
-    <dimen name="lessons_info_extended_text_size">30sp</dimen>
     <dimen name="lessons_letter_vertical_bias">0.0</dimen>
     <dimen name="lessons_dots_info_width">394dp</dimen>
     <dimen name="lessons_dots_info_height">175dp</dimen>
